feat: move service to folder
